FIX: test_basic, 64 bit truncation - use "long long" for a 64 bit type on      gcc/Linux.
    I used uint64_t, but I should have used "long long" instead.

1.1 -> trunk.
----------------------------------------------------------
e --include-merges or -n0 to see merged revisions.

+++ tests/test_basic.c
@@ -132,7 +132,7 @@ FCT_BGN()
 #if defined(WIN32)
             __int64 sixfourval =0;
 #else
-            uint64_t  sixfourval =0;
+            unsigned long long  sixfourval =0;
 #endif
             /* G++ compilers want to see the LLU (long long unsigned)
             tag at the of the variable, Microsoft compilers are sad when
